+++ californium-core/pom.xml
@@ -29,6 +29,10 @@
 			<groupId>${project.groupId}</groupId>
 			<artifactId>element-connector</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>element-connector-tcp</artifactId>
+		</dependency>
 	</dependencies>
 
 	<profiles>
