+++ element-connector-tcp/src/main/java/org/eclipse/californium/elements/tcp/CloseOnErrorHandler.java
