+++ element-connector-tcp/src/main/java/org/eclipse/californium/elements/tcp/CloseOnIdleHandler.java
