+++ element-connector-tcp/src/main/resources/about.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1" />
+<title>About</title>
+</head>
+<body lang="EN-US">
+<h2>About This Content</h2>
+ 
+<p><em>November 16, 2015</em></p>	
+<h3>License</h3>
+
+<p>The Eclipse Foundation makes available all content in this plug-in ("Content").  Unless otherwise 
+indicated below, the Content is provided to you under the terms and conditions of the
+Eclipse Public License Version 1.0 ("EPL") and Eclipse Distribution License Version 1.0 ("EDL").
+A copy of the EPL is available at 
+<a href="http://www.eclipse.org/legal/epl-v10.html">http://www.eclipse.org/legal/epl-v10.html</a> 
+and a copy of the EDL is available at 
+<a href="http://www.eclipse.org/org/documents/edl-v10.html">http://www.eclipse.org/org/documents/edl-v10.html</a>. 
+For purposes of the EPL, "Program" will mean the Content.</p>
+
+<p>If you did not receive this Content directly from the Eclipse Foundation, the Content is being
+redistributed by another party ("Redistributor") and different terms and conditions may apply to your
+use of any object code in the Content. Check the Redistributor's license that was provided with the Content.
+If no such license exists, contact the Redistributor. Unless otherwise indicated below, the terms and conditions
+of the EPL and EDL still apply to any source code in the Content and such source code may be obtained at
+<a href="http://www.eclipse.org/">http://www.eclipse.org</a>.</p>
+
+<h3>Third Party Content</h3>
+
+<p>The Content includes items that have been sourced from third parties as set out below.
+If you did not receive this Content directly from the Eclipse Foundation, the following is provided
+for informational purposes only, and you should look to the Redistributor's license for terms and
+conditions of use.</p>
+
+<h5>Heinz M. Kabutz' Striped Executor Service</h5>
+
+<p>This product includes the Striped Executor Service implementation developed by Heinz M. Kabutz.
+(<a href="http://www.javaspecialists.eu/archive/Issue206.html">206 Striped Executor Service</a>).</p>
+
+<p>Your use of the <em>Striped Executor Service</em> is subject to the terms and conditions of the Apache Software License 2.0.
+A copy of the license is contained in the file <a href="LICENSE-2.0.txt">LICENSE-2.0.txt</a> and is also available at
+<a href="http://www.apache.org/licenses/LICENSE-2.0.html">http://www.apache.org/licenses/LICENSE-2.0.html</a>.</p>
+
+<p>The source code is available from
+<a href="https://github.com/kabutz/javaspecialists">Heinz M. Kabutz' GitHub repository</a>.</p>
+</body>
+</html>
