+++ pom.xml
@@ -163,6 +163,11 @@
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
+				<groupId>${project.groupId}</groupId>
+				<artifactId>element-connector-tcp</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
 				<groupId>org.mockito</groupId>
 				<artifactId>mockito-all</artifactId>
 				<version>1.9.5</version>
@@ -174,6 +179,7 @@
 	<modules>
 		<module>element-connector</module>
 		<module>scandium-core</module>
+		<module>element-connector-tcp</module>
 		<module>californium-core</module>
 		<module>californium-proxy</module>
 		<module>californium-osgi</module>
